requirements.txt
pyautogui
Pillow
keyboard
reportlab
